refactor Vessel search and use useClickOutside hook
