+++ frontend/src/containers/VesselVisibility.js
@@ -19,6 +19,7 @@ import unselectVessel from '../domain/use_cases/unselectVessel'
 import ShowVesselEstimatedPositions from '../components/vessel_visibility/ShowVesselEstimatedPositions'
 import { MapComponentStyle } from '../components/commonStyles/MapComponent.style'
 import { MapButtonStyle } from '../components/commonStyles/MapButton.style'
+import { useClickOutsideComponent } from '../hooks/useClickOutside'
 
 const VesselVisibility = () => {
   const dispatch = useDispatch()
@@ -31,23 +32,14 @@ const VesselVisibility = () => {
   const { healthcheckTextWarning } = useSelector(state => state.global)
 
   const [vesselVisibilityBoxIsOpen, setVesselVisibilityBoxIsOpen] = useState(false)
-
   const wrapperRef = useRef(null)
+  const clickedOutsideComponent = useClickOutsideComponent(wrapperRef)
 
   useEffect(() => {
-    function handleClickOutside (event) {
-      if (wrapperRef.current && !wrapperRef.current.contains(event.target)) {
-        setVesselVisibilityBoxIsOpen(false)
-      }
+    if (clickedOutsideComponent) {
+      setVesselVisibilityBoxIsOpen(false)
     }
-
-    // Bind the event listener
-    document.addEventListener('mousedown', handleClickOutside)
-    return () => {
-      // Unbind the event listener on clean up
-      document.removeEventListener('mousedown', handleClickOutside)
-    }
-  }, [wrapperRef])
+  }, [clickedOutsideComponent])
 
   useEffect(() => {
     if (vesselVisibilityBoxIsOpen === true) {
